+++ src/app/shared/shared.module.ts
@@ -1,10 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { BackgroundColorDirective } from './directives/background-color.directive';
+
 @NgModule({
-  declarations: [],
+  declarations: [
+    BackgroundColorDirective
+  ],
   imports: [
     CommonModule
+  ],
+  exports: [
+    BackgroundColorDirective
   ]
 })
 export class SharedModule { }